Refactor module structure
Moved `__main__.py` to `main.py` and updated import paths accordingly. This
change improves the project's structure by separating the main entry point
from the module's implementation, making it more maintainable and easier to
understand. The refactoring involved updating import statements in both the
test files and the `pyproject.toml` configuration to reflect the new module
path.
